Beta 2: 1. Tinyxml as default XML implementation 2. New parameter code added 3. New InstrConfig class added

+++ Src/Common/SynthString.cpp
@@ -43,12 +43,12 @@ int bsString::SetLen(int newLen)
 		curLen = strlen(theStr);
 	else
 	{
-		if (newLen >= maxLen)
-			Allocate(newLen);
+		if ((size_t)newLen >= maxLen)
+			Allocate((size_t)newLen);
 		curLen = newLen;
 		theStr[curLen] = 0;
 	}
-	return curLen;
+	return (int) curLen;
 }
 
 bsString& bsString::Assign(const char *s)
